+++ tests/integration/PayPalAccountTest.php
@@ -30,6 +30,7 @@ class PayPalAccountTest extends Setup
         $this->assertSame('jane.doe@example.com', $foundPayPalAccount->email);
         $this->assertSame($paymentMethodToken, $foundPayPalAccount->token);
         $this->assertNotNull($foundPayPalAccount->imageUrl);
+        $this->assertNull($foundPayPalAccount->revokedAt);
     }
 
     public function testGatewayFind()