+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+* Add `revokedAt` field to `PayPalAccount`
+* Add support for `PAYMENT_METHOD_REVOKED_BY_CUSTOMER` webhook
+
 ## 4.0.0
 * Add support for PHP 7 (Thanks to @briandotdev)
 * Remove support for HHVM and PHP 5